+++ cat-consumer/src/main/java/org/unidal/cat/plugin/transaction/TransactionPipeline.java
@@ -1,6 +1,5 @@
 package org.unidal.cat.plugin.transaction;
 
-import com.dianping.cat.Cat;
 import com.dianping.cat.Constants;
 import org.unidal.cat.spi.Report;
 import org.unidal.cat.spi.ReportManager;
@@ -10,6 +9,7 @@ import org.unidal.cat.spi.analysis.pipeline.AbstractPipeline;
 import org.unidal.cat.spi.analysis.pipeline.Pipeline;
 import org.unidal.cat.spi.report.ReportDelegate;
 import org.unidal.cat.spi.report.internals.ReportDelegateManager;
+import org.unidal.lookup.annotation.Inject;
 import org.unidal.lookup.annotation.Named;
 
 import java.io.IOException;
@@ -19,33 +19,32 @@ import java.util.Map;
 
 @Named(type = Pipeline.class, value = TransactionConstants.NAME, instantiationStrategy = Named.PER_LOOKUP)
 public class TransactionPipeline extends AbstractPipeline {
+    @Inject
+    private ReportManagerManager m_rmm;
+
+    @Inject
+    private ReportDelegateManager m_rdg;
+
     @Override
-    protected void beforeCheckpoint() {
-        ReportManagerManager rmm = super.lookup(ReportManagerManager.class);
-        ReportManager<Report> reportManager = rmm.getReportManager(getName());
+    protected void beforeCheckpoint() throws IOException {
+        ReportManager<Report> manager = m_rmm.getReportManager(getName());
 
-        ReportDelegateManager rdg = super.lookup(ReportDelegateManager.class);
-        ReportDelegate<Report> reportDelegate = rdg.getDelegate(getName());
+        ReportDelegate<Report> delegate = m_rdg.getDelegate(getName());
 
+        List<Map<String, Report>> reportMapList = manager.getLocalReports(ReportPeriod.HOUR, getHour());
 
-        try {
-            List<Map<String, Report>> reportMapList = reportManager.getLocalReports(ReportPeriod.HOUR, getHour());
-            if(reportMapList.size() > 0){
-                List<Report> reportList = new ArrayList<Report>();
+        if (reportMapList.size() > 0) {
+            List<Report> reportList = new ArrayList<Report>();
 
-                for (Map<String, Report> map : reportMapList) {
-                    for (Report report : map.values()) {
-                        reportList.add(report);
-                    }
-                }
-
-                Report allReport = reportDelegate.makeAllReport(ReportPeriod.HOUR, reportList);
-
-                Map<String, Report> map = reportMapList.get(0);
-                map.put(Constants.ALL, allReport);
+            for (Map<String, Report> map : reportMapList) {
+                reportList.addAll(map.values());
             }
-        } catch (IOException e) {
-            Cat.logError(e);
+
+            Report allReport = delegate.makeAll(ReportPeriod.HOUR, reportList);
+
+            Map<String, Report> map = reportMapList.get(0);
+
+            map.put(Constants.ALL, allReport);
         }
     }
 }
