+++ cat-consumer/src/main/java/org/unidal/cat/plugin/transaction/TransactionReportDelegate.java
@@ -28,8 +28,8 @@ public class TransactionReportDelegate implements ReportDelegate<TransactionRepo
 	}
 
 	@Override
-	public TransactionReport makeAllReport(ReportPeriod period, Collection<TransactionReport> reports) {
-		return m_aggregator.makeAllReport(period, reports);
+	public TransactionReport makeAll(ReportPeriod period, Collection<TransactionReport> reports) {
+		return m_aggregator.makeAll(period, reports);
 	}
 
 	@Override
