+++ cat-consumer/src/main/java/org/unidal/cat/plugin/transaction/filter/TransactionAllNameFilter.java
@@ -2,6 +2,7 @@ package org.unidal.cat.plugin.transaction.filter;
 
 import com.dianping.cat.Constants;
 import com.dianping.cat.consumer.transaction.model.entity.Machine;
+import com.dianping.cat.consumer.transaction.model.entity.TransactionName;
 import com.dianping.cat.consumer.transaction.model.entity.TransactionReport;
 import com.dianping.cat.consumer.transaction.model.entity.TransactionType;
 import com.dianping.cat.consumer.transaction.model.transform.BaseVisitor;
@@ -11,12 +12,9 @@ import org.unidal.cat.spi.report.ReportFilter;
 import org.unidal.lookup.annotation.Inject;
 import org.unidal.lookup.annotation.Named;
 
-import java.util.ArrayList;
-import java.util.Collection;
-
-@Named(type = ReportFilter.class, value = TransactionConstants.NAME + ":" + TransactionAllDomainTypeFilter.ID)
-public class TransactionAllDomainTypeFilter implements ReportFilter<TransactionReport> {
-    public static final String ID = "all-domain-type";
+@Named(type = ReportFilter.class, value = TransactionConstants.NAME + ":" + TransactionAllNameFilter.ID)
+public class TransactionAllNameFilter implements ReportFilter<TransactionReport> {
+    public static final String ID = "all-name";
 
     @Inject
     private TransactionReportHelper m_helper;
@@ -33,7 +31,8 @@ public class TransactionAllDomainTypeFilter implements ReportFilter<TransactionR
 
     @Override
     public TransactionReport screen(RemoteContext ctx, TransactionReport report) {
-        TypeScreener visitor = new TypeScreener(report.getDomain());
+        String type = ctx.getProperty("type", null);
+        NameScreener visitor = new NameScreener(report.getDomain(), type);
 
         report.accept(visitor);
         return visitor.getReport();
@@ -41,17 +40,20 @@ public class TransactionAllDomainTypeFilter implements ReportFilter<TransactionR
 
     @Override
     public void tailor(RemoteContext ctx, TransactionReport report) {
+        String type = ctx.getProperty("type", null);
         String ip = ctx.getProperty("ip", null);
-        TypeTailor visitor = new TypeTailor(ip);
+        NameTailor visitor = new NameTailor(type, ip);
 
         report.accept(visitor);
     }
 
-    private class TypeScreener extends BaseVisitor {
+    private class NameScreener extends BaseVisitor {
+        private String m_typeName;
 
         private TransactionHolder m_holder = new TransactionHolder();
 
-        public TypeScreener(String domain) {
+        public NameScreener(String domain, String type) {
+            m_typeName = type;
             m_holder.setReport(new TransactionReport(domain));
         }
 
@@ -62,17 +64,22 @@ public class TransactionAllDomainTypeFilter implements ReportFilter<TransactionR
         @Override
         public void visitMachine(Machine machine) {
             Machine m = m_holder.getMachine();
+            TransactionType t = m.findOrCreateType(m_typeName);
+            TransactionType type = machine.findType(m_typeName);
 
             m_helper.mergeMachine(m, machine);
+            m_holder.setType(t);
 
-            Collection<TransactionType> types = new ArrayList<TransactionType>(machine.getTypes().values());
-
-            for (TransactionType type : types) {
-                TransactionType t = m.findOrCreateType(type.getId());
-
-                m_holder.setType(t);
+            if (type != null) {
                 visitType(type);
             }
+        }
+
+        @Override
+        public void visitName(TransactionName name) {
+            TransactionName n = m_holder.getName();
+
+            m_helper.mergeName(n, name);
         }
 
         @Override
@@ -82,11 +89,10 @@ public class TransactionAllDomainTypeFilter implements ReportFilter<TransactionR
             m_helper.mergeReport(r, report);
 
             Machine m = r.findOrCreateMachine(Constants.ALL);
-            Collection<Machine> machines = new ArrayList<Machine>(report.getMachines().values());
 
             m_holder.setMachine(m);
 
-            for (Machine machine : machines) {
+            for (Machine machine : report.getMachines().values()) {
                 visitMachine(machine);
             }
         }
@@ -96,16 +102,55 @@ public class TransactionAllDomainTypeFilter implements ReportFilter<TransactionR
             TransactionType t = m_holder.getType();
 
             m_helper.mergeType(t, type);
+
+            for (TransactionName name : type.getNames().values()) {
+                TransactionName n = t.findOrCreateName(name.getId());
+
+                m_holder.setName(n);
+                visitName(name);
+            }
         }
     }
 
-    private class TypeTailor extends BaseVisitor {
+    private class NameTailor extends BaseVisitor {
+        private String m_typeName;
+
         private String m_ip;
 
         private Machine m_machine;
 
-        public TypeTailor(String ip) {
+        private TransactionType m_type;
+
+        public NameTailor(String type, String ip) {
+            m_typeName = type;
             m_ip = ip;
+        }
+
+        @Override
+        public void visitMachine(Machine machine) {
+            TransactionType type = machine.findType(m_typeName);
+
+            machine.getTypes().clear();
+
+            if (type != null) {
+                m_type = m_machine.findOrCreateType(type.getId());
+
+                m_helper.mergeType(m_type, type);
+                machine.addType(type);
+            }
+
+            super.visitMachine(machine);
+        }
+
+        @Override
+        public void visitName(TransactionName name) {
+            name.getRanges().clear();
+            name.getDurations().clear();
+            name.getAllDurations().clear();
+
+            TransactionName n = m_type.findOrCreateName(name.getId());
+
+            m_helper.mergeName(n, name);
         }
 
         @Override
@@ -130,13 +175,10 @@ public class TransactionAllDomainTypeFilter implements ReportFilter<TransactionR
 
         @Override
         public void visitType(TransactionType type) {
-            type.getNames().clear();
             type.getRange2s().clear();
             type.getAllDurations().clear();
 
-            TransactionType t = m_machine.findOrCreateType(type.getId());
-
-            m_helper.mergeType(t, type);
+            super.visitType(type);
         }
     }
 }
