+++ cat-consumer/src/main/java/org/unidal/cat/plugin/transaction/filter/TransactionAllTypeFilter.java
@@ -2,7 +2,6 @@ package org.unidal.cat.plugin.transaction.filter;
 
 import com.dianping.cat.Constants;
 import com.dianping.cat.consumer.transaction.model.entity.Machine;
-import com.dianping.cat.consumer.transaction.model.entity.TransactionName;
 import com.dianping.cat.consumer.transaction.model.entity.TransactionReport;
 import com.dianping.cat.consumer.transaction.model.entity.TransactionType;
 import com.dianping.cat.consumer.transaction.model.transform.BaseVisitor;
@@ -12,9 +11,12 @@ import org.unidal.cat.spi.report.ReportFilter;
 import org.unidal.lookup.annotation.Inject;
 import org.unidal.lookup.annotation.Named;
 
-@Named(type = ReportFilter.class, value = TransactionConstants.NAME + ":" + TransactionAllDomainNameFilter.ID)
-public class TransactionAllDomainNameFilter implements ReportFilter<TransactionReport> {
-    public static final String ID = "all-domain-name";
+import java.util.ArrayList;
+import java.util.Collection;
+
+@Named(type = ReportFilter.class, value = TransactionConstants.NAME + ":" + TransactionAllTypeFilter.ID)
+public class TransactionAllTypeFilter implements ReportFilter<TransactionReport> {
+    public static final String ID = "all-type";
 
     @Inject
     private TransactionReportHelper m_helper;
@@ -31,8 +33,7 @@ public class TransactionAllDomainNameFilter implements ReportFilter<TransactionR
 
     @Override
     public TransactionReport screen(RemoteContext ctx, TransactionReport report) {
-        String type = ctx.getProperty("type", null);
-        NameScreener visitor = new NameScreener(report.getDomain(), type);
+        TypeScreener visitor = new TypeScreener(report.getDomain());
 
         report.accept(visitor);
         return visitor.getReport();
@@ -40,20 +41,17 @@ public class TransactionAllDomainNameFilter implements ReportFilter<TransactionR
 
     @Override
     public void tailor(RemoteContext ctx, TransactionReport report) {
-        String type = ctx.getProperty("type", null);
         String ip = ctx.getProperty("ip", null);
-        NameTailor visitor = new NameTailor(type, ip);
+        TypeTailor visitor = new TypeTailor(ip);
 
         report.accept(visitor);
     }
 
-    private class NameScreener extends BaseVisitor {
-        private String m_typeName;
+    private class TypeScreener extends BaseVisitor {
 
         private TransactionHolder m_holder = new TransactionHolder();
 
-        public NameScreener(String domain, String type) {
-            m_typeName = type;
+        public TypeScreener(String domain) {
             m_holder.setReport(new TransactionReport(domain));
         }
 
@@ -64,22 +62,17 @@ public class TransactionAllDomainNameFilter implements ReportFilter<TransactionR
         @Override
         public void visitMachine(Machine machine) {
             Machine m = m_holder.getMachine();
-            TransactionType t = m.findOrCreateType(m_typeName);
-            TransactionType type = machine.findType(m_typeName);
 
             m_helper.mergeMachine(m, machine);
-            m_holder.setType(t);
 
-            if (type != null) {
+            Collection<TransactionType> types = new ArrayList<TransactionType>(machine.getTypes().values());
+
+            for (TransactionType type : types) {
+                TransactionType t = m.findOrCreateType(type.getId());
+
+                m_holder.setType(t);
                 visitType(type);
             }
-        }
-
-        @Override
-        public void visitName(TransactionName name) {
-            TransactionName n = m_holder.getName();
-
-            m_helper.mergeName(n, name);
         }
 
         @Override
@@ -89,10 +82,11 @@ public class TransactionAllDomainNameFilter implements ReportFilter<TransactionR
             m_helper.mergeReport(r, report);
 
             Machine m = r.findOrCreateMachine(Constants.ALL);
+            Collection<Machine> machines = new ArrayList<Machine>(report.getMachines().values());
 
             m_holder.setMachine(m);
 
-            for (Machine machine : report.getMachines().values()) {
+            for (Machine machine : machines) {
                 visitMachine(machine);
             }
         }
@@ -102,55 +96,16 @@ public class TransactionAllDomainNameFilter implements ReportFilter<TransactionR
             TransactionType t = m_holder.getType();
 
             m_helper.mergeType(t, type);
-
-            for (TransactionName name : type.getNames().values()) {
-                TransactionName n = t.findOrCreateName(name.getId());
-
-                m_holder.setName(n);
-                visitName(name);
-            }
         }
     }
 
-    private class NameTailor extends BaseVisitor {
-        private String m_typeName;
-
+    private class TypeTailor extends BaseVisitor {
         private String m_ip;
 
         private Machine m_machine;
 
-        private TransactionType m_type;
-
-        public NameTailor(String type, String ip) {
-            m_typeName = type;
+        public TypeTailor(String ip) {
             m_ip = ip;
-        }
-
-        @Override
-        public void visitMachine(Machine machine) {
-            TransactionType type = machine.findType(m_typeName);
-
-            machine.getTypes().clear();
-
-            if (type != null) {
-                m_type = m_machine.findOrCreateType(type.getId());
-
-                m_helper.mergeType(m_type, type);
-                machine.addType(type);
-            }
-
-            super.visitMachine(machine);
-        }
-
-        @Override
-        public void visitName(TransactionName name) {
-            name.getRanges().clear();
-            name.getDurations().clear();
-            name.getAllDurations().clear();
-
-            TransactionName n = m_type.findOrCreateName(name.getId());
-
-            m_helper.mergeName(n, name);
         }
 
         @Override
@@ -175,10 +130,13 @@ public class TransactionAllDomainNameFilter implements ReportFilter<TransactionR
 
         @Override
         public void visitType(TransactionType type) {
+            type.getNames().clear();
             type.getRange2s().clear();
             type.getAllDurations().clear();
 
-            super.visitType(type);
+            TransactionType t = m_machine.findOrCreateType(type.getId());
+
+            m_helper.mergeType(t, type);
         }
     }
 }
