+++ cat-core/src/main/java/org/unidal/cat/spi/report/ReportDelegate.java
@@ -12,7 +12,7 @@ import org.unidal.cat.spi.ReportPeriod;
 public interface ReportDelegate<T extends Report> {
 	public T aggregate(ReportPeriod period, Collection<T> reports);
 
-	public T makeAllReport(ReportPeriod period, Collection<T> reports);
+	public T makeAll(ReportPeriod period, Collection<T> reports);
 
 	public String buildXml(T report);
